# FreeTimeApp/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "tiempo-libre",
    "version": "1.0"
  },
  "paths": {
...
    "/api/validacion-porcentaje/validar-grupo/{grupoId}": {
      "get": {
        "tags": [
          "ValidacionPorcentaje"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ausenciasSolicitadas",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/validacion-porcentaje/estado-grupo/{grupoId}": {
      "get": {
        "tags": [
          "ValidacionPorcentaje"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/validacion-porcentaje/calcular-minimo-empleados": {
      "get": {
        "tags": [
          "ValidacionPorcentaje"
        ],
        "parameters": [
          {
            "name": "porcentaje",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
...
}